moved remaining grobid trainer to cli module
