+++ sciencebeam_trainer_delft/sequence_labelling/tools/grobid_trainer/utils.py
@@ -1,3 +1,4 @@
+# mostly copied from https://github.com/kermitt2/delft/blob/master/grobidTagger.py
 import logging
 import time
 import tempfile
